Two sample apps.
"books" is a native version of the HTML5 example in Tim Anglade's HTML5 class.
"browser" is a simple browser of Usergrid collections.

This commit also includes a minor change to UGConnection to fix
crashes in the sample apps that would occur when users attempted
to sign in with no username or password specified.